Change input to int from float to match output

+++ prac_04/list_exercises.py
@@ -5,7 +5,7 @@ You can use the functions min, max, sum and len, and you can use the append meth
 numbers = []
 
 for i in range(5):
-    number = float(input("Number: "))
+    number = int(input("Number: "))
     numbers.append(number)
 
 print(f"The first number is {numbers[0]}")
